Move logger to shared folder
